LayoutTests/imported/w3c/web-platform-tests/trusted-types/HTMLScriptElement-src-expected.txt

CONSOLE MESSAGE: This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

PASS Setting .src to a plain string should throw an exception and not modify the script state, on an unconnected script element.
PASS Setting .src to a TrustedScriptURL should work and should execute the referenced script instead of the script body, on an unconnected script element.
PASS Setting .src to a plain string should throw an exception and not modify the script state, on a connected script element.
PASS Setting .src to a TrustedScriptURL should work and should execute the referenced script instead of the script body, on a onnected script element.
